spikes/pyproject.toml
[project]
name = "construct-spikes"
version = "0.1.0"
description = "Spike experiments for Construct browser automation"
requires-python = ">=3.11"
dependencies = [
    "workflow-use",
    "browser-use",
    "openai",
    "python-dotenv",
    "playwright",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
